+++ src/components/renderer/LookupNotification.tsx
@@ -344,8 +344,9 @@ function LookupNotificationComponent({
       isLoading,
       onLookup: triggerLookup,
       allowLookupOnEmptyValue: runLookupOnClear,
+      isLookingUp,
     }),
-    [isDisabled, isLoading, runLookupOnClear, triggerLookup],
+    [isDisabled, isLoading, runLookupOnClear, triggerLookup, isLookingUp],
   )
 
   return (