+++ src/form-elements/FormElementSelect.tsx
@@ -12,6 +12,7 @@ import {
   IsDirtyProps,
   UpdateFormElementsHandler,
 } from '../types/form'
+import { LookupNotificationContext } from '../hooks/useLookupNotification'
 
 type Props = {
   id: string
@@ -51,6 +52,10 @@ function FormElementSelect({
     if (typeof value === 'string') return [value]
     return []
   }, [value])
+
+  const { isLookingUp } = React.useContext(LookupNotificationContext)
+  const isDisplayingValidationMessage =
+    (isDirty || displayValidationMessage) && !!validationMessage && !isLookingUp
 
   return (
     <div className="cypress-select-element">
@@ -139,7 +144,7 @@ function FormElementSelect({
             </div>
           )}
 
-          {(isDirty || displayValidationMessage) && !!validationMessage && (
+          {isDisplayingValidationMessage && (
             <div role="alert" className="has-margin-top-8">
               <div className="has-text-danger ob-error__text cypress-validation-message">
                 {validationMessage}