+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Use lookup button label for required validation message if configured
 
+### Fixed
+
+- 'Lookup is Required' message showing while lookup is running
+
 ## [5.2.0] - 2023-10-25
 
 ### Added